- Default Setting now "Off" - seperate Update Thread - moved updateManager to lib folder
