+++ resources/lib/updateManager.py
